Rivisi Bug #1

+++ app/src/main/java/com/ikm/inventoryqrscanner/activity/ProductActivity.kt
@@ -75,7 +75,7 @@ class ProductActivity : BaseActivity() {
             number = document["number"].toString(),
             product = document["product"].toString(),
             expDate = document["expDate"] as? Timestamp,
-            amount = document["amount"].toString().toInt(),
+            amount = document["amount"].toString(),
             type = document["type"].toString(),
             location = document["location"].toString(),
             condition = document["condition"].toString(),
